src/com/ppc/service/FileProcessorTest.java
package com.ppc.service;
/*
import static org.mockito.Mockito.*;
import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;
import org.mockito.*;
import java.io.*;
import java.util.*;*/

public class FileProcessorTest {

	/* @Mock
    private BufferedReader mockedReader;
    
    @Mock
    private FileReader mockedFileReader;

    private FileProcessor fileProcessor;

    @Before
    public void setUp() throws Exception {
        // Initialize the mocks
        MockitoAnnotations.initMocks(this);
        
        // Mocking the file reader and buffered reader behavior
        when(mockedFileReader.read()).thenReturn(-1);  // Simulate end of file
        when(mockedReader.readLine()).thenReturn("Header")
                                      .thenReturn("1,EMP001,ONBOARD,2024-01-01,Developer,New Employee")
                                      .thenReturn("2,EMP002,SALARY,5000.00,2024-01-01")
                                      .thenReturn(null);  // End of file
        
        // Instantiate FileProcessor
        fileProcessor = new FileProcessor();
    }

    @Test
    public void testProcessFile() throws IOException {
        // Assume processFile is tested with mocked input data
        fileProcessor.processFile("dummy/path/to/file.csv");

        // Validate that the file was processed
        assertEquals("Total employees should be 2 after processing", 2, fileProcessor.getTotalEmployees());

        // Check if a specific employee was onboarded
        List<IEmployeeDetails> newEmployees = fileProcessor.getNewEmployeesOfMonth(LocalDate.of(2024, 1, 1));
        assertEquals("There should be 1 new employee onboarded in January 2024", 1, newEmployees.size());

        // Check if the employee is correctly onboarded
        IEmployeeDetails onboardedEmployee = newEmployees.get(0);
        assertEquals("Employee ID should be EMP001", "EMP001", onboardedEmployee.getEmpId());
        assertEquals("Employee FirstName should be New Employee", "New Employee", onboardedEmployee.getFirstName());

        // Check if the salary event was processed correctly
        IMonthlyReportFormat salaryReport = fileProcessor.getTotalSalaryOfMonth(LocalDate.of(2024, 1, 1));
        assertEquals("Total salary for January 2024 should be 5000.00", 5000.00, salaryReport.getAmount(), 0.01);
    }

    @Test
    public void testProcessFileWithInvalidData() throws IOException {
        // Simulating invalid data where fields are insufficient (less than 6 fields)
        when(mockedReader.readLine()).thenReturn("Invalid data line").thenReturn(null);

        // Running the file processor with invalid data
        fileProcessor.processFile("dummy/path/to/invalid_file.csv");

        // Check the size should remain zero since invalid data was processed
        assertEquals("Total employees should be 0 due to invalid data", 0, fileProcessor.getTotalEmployees());
    }

    @Test
    public void testProcessExitEvent() throws IOException {
        // Mock an employee exit event
        when(mockedReader.readLine()).thenReturn("3,EMP001,EXIT,2024-12-31").thenReturn(null);
        
        // Process exit event
        fileProcessor.processFile("dummy/path/to/file.csv");
        
        // Validate employee exit
        Employee employee = fileProcessor.employees.get("EMP001");
        assertNotNull("Employee EMP001 should exist", employee);
        assertEquals("Exit date should be 2024-12-31", Helper.parseDate("2024-12-31"), employee.getexitDate());
    }

    @Test
    public void testGetAnnualReport() throws IOException {
        // Assuming some events were processed in a previous test
        fileProcessor.processFile("dummy/path/to/file.csv");

        // Get the annual report for a specific year
        List<ICompleteEvent> annualReport = fileProcessor.getAnnualReport(LocalDate.of(2024, 1, 1));
        
        // Validate the report contains the expected number of events for the year 2024
        assertEquals("There should be 2 events in the annual report for 2024", 2, annualReport.size());
    }

    @Test
    public void testGetEmployeeMonthlyReport() throws IOException {
        // Simulate processing of a salary event
        fileProcessor.processFile("dummy/path/to/file.csv");

        // Get the monthly employee report
        List<ITotalMonthlyReportFormat> report = fileProcessor.getEmployeeMonthlyReport(LocalDate.of(2024, 1, 1));

        // Validate the report for specific employee details
        assertEquals("There should be one entry in the monthly report", 1, report.size());
        assertEquals("Employee ID in the report should be EMP002", "EMP002", report.get(0).getEmpId());
        assertEquals("Employee FirstName should be Developer", "Developer", report.get(0).getFirstName());
    }*/
}
